ref(onboarding): Split python onboarding docs
