+++ static/app/gettingStartedDocs/python/chalice.tsx
@@ -4,11 +4,9 @@ import {
   type DocsParams,
   type OnboardingConfig,
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
-import {
-  agentMonitoringOnboarding,
-  crashReportOnboardingPython,
-  mcpOnboarding,
-} from 'sentry/gettingStartedDocs/python/python';
+import {agentMonitoring} from 'sentry/gettingStartedDocs/python/python/agentMonitoring';
+import {crashReport} from 'sentry/gettingStartedDocs/python/python/crashReport';
+import {mcp} from 'sentry/gettingStartedDocs/python/python/mcp';
 import {t, tct} from 'sentry/locale';
 import {
   alternativeProfilingConfiguration,
@@ -149,9 +147,9 @@ const logsOnboarding = getPythonLogsOnboarding({
 const docs: Docs = {
   onboarding,
   profilingOnboarding: getPythonProfilingOnboarding({basePackage: 'sentry-sdk[chalice]'}),
-  crashReportOnboarding: crashReportOnboardingPython,
-  agentMonitoringOnboarding,
-  mcpOnboarding,
+  crashReportOnboarding: crashReport,
+  agentMonitoringOnboarding: agentMonitoring,
+  mcpOnboarding: mcp,
   logsOnboarding,
 };
