+++ static/app/gettingStartedDocs/python/gcpfunctions.tsx
@@ -5,11 +5,9 @@ import {
   type DocsParams,
   type OnboardingConfig,
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
-import {
-  agentMonitoringOnboarding,
-  crashReportOnboardingPython,
-  mcpOnboarding,
-} from 'sentry/gettingStartedDocs/python/python';
+import {agentMonitoring} from 'sentry/gettingStartedDocs/python/python/agentMonitoring';
+import {crashReport} from 'sentry/gettingStartedDocs/python/python/crashReport';
+import {mcp} from 'sentry/gettingStartedDocs/python/python/mcp';
 import {t, tct} from 'sentry/locale';
 import {
   alternativeProfilingConfiguration,
@@ -197,10 +195,10 @@ const logsOnboarding = getPythonLogsOnboarding();
 
 const docs: Docs = {
   onboarding,
-  crashReportOnboarding: crashReportOnboardingPython,
+  crashReportOnboarding: crashReport,
   profilingOnboarding: getPythonProfilingOnboarding(),
-  agentMonitoringOnboarding,
-  mcpOnboarding,
+  agentMonitoringOnboarding: agentMonitoring,
+  mcpOnboarding: mcp,
   logsOnboarding,
 };
