+++ static/app/gettingStartedDocs/python/python/onboarding.spec.tsx
@@ -6,7 +6,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './python';
+import docs from '.';
 
 describe('python onboarding docs', () => {
   it('renders doc correctly', () => {
