+++ static/app/gettingStartedDocs/python/serverless.tsx
@@ -7,10 +7,8 @@ import {
   type DocsParams,
   type OnboardingConfig,
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
-import {
-  agentMonitoringOnboarding,
-  crashReportOnboardingPython,
-} from 'sentry/gettingStartedDocs/python/python';
+import {agentMonitoring} from 'sentry/gettingStartedDocs/python/python/agentMonitoring';
+import {crashReport} from 'sentry/gettingStartedDocs/python/python/crashReport';
 import {t, tct} from 'sentry/locale';
 import {
   alternativeProfilingConfiguration,
@@ -175,8 +173,8 @@ const onboarding: OnboardingConfig = {
 const docs: Docs = {
   onboarding,
   profilingOnboarding: getPythonProfilingOnboarding(),
-  crashReportOnboarding: crashReportOnboardingPython,
-  agentMonitoringOnboarding,
+  crashReportOnboarding: crashReport,
+  agentMonitoringOnboarding: agentMonitoring,
 };
 
 export default docs;
