+++ static/app/gettingStartedDocs/python/tryton.tsx
@@ -5,11 +5,9 @@ import type {
   OnboardingConfig,
 } from 'sentry/components/onboarding/gettingStartedDoc/types';
 import {StepType} from 'sentry/components/onboarding/gettingStartedDoc/types';
-import {
-  agentMonitoringOnboarding,
-  crashReportOnboardingPython,
-  mcpOnboarding,
-} from 'sentry/gettingStartedDocs/python/python';
+import {agentMonitoring} from 'sentry/gettingStartedDocs/python/python/agentMonitoring';
+import {crashReport} from 'sentry/gettingStartedDocs/python/python/crashReport';
+import {mcp} from 'sentry/gettingStartedDocs/python/python/mcp';
 import {t, tct} from 'sentry/locale';
 import {
   alternativeProfilingConfiguration,
@@ -169,9 +167,9 @@ const profilingOnboarding: OnboardingConfig = {
 const docs: Docs = {
   onboarding,
   profilingOnboarding,
-  crashReportOnboarding: crashReportOnboardingPython,
-  agentMonitoringOnboarding,
-  mcpOnboarding,
+  crashReportOnboarding: crashReport,
+  agentMonitoringOnboarding: agentMonitoring,
+  mcpOnboarding: mcp,
   logsOnboarding,
 };
